Introduce AdSense module asset based on API from #1321.
