+++ webpack.config.js
@@ -92,7 +92,8 @@ const webpackConfig = ( mode ) => {
 				'googlesitekit-api': './assets/js/googlesitekit-api.js',
 				'googlesitekit-data': './assets/js/googlesitekit-data.js',
 				'googlesitekit-datastore-site': './assets/js/googlesitekit-datastore-site.js',
-				'googlesitekit-modules': './assets/js/googlesitekit-modules.js', // TODO: Add external following 1162.
+				'googlesitekit-modules': './assets/js/googlesitekit-modules.js',
+				'googlesitekit-modules-adsense': './assets/js/googlesitekit-modules-adsense.js',
 				// Old Modules
 				'googlesitekit-activation': './assets/js/googlesitekit-activation.js',
 				'googlesitekit-settings': './assets/js/googlesitekit-settings.js',
